requirements.txt
PySide6>=6.5
python-docx>=1.0.0
PyMuPDF>=1.24.0
PyPDF2>=3.0.0
deep-translator>=1.11.4
requests>=2.31.0